Moved World-related stuff into a separate module
